+++ modules/firefox.nix
@@ -1,0 +1,11 @@
+{
+  hmModule = { ... }: {
+    programs.firefox.enable = true;
+    programs.firefox.profiles.default = {
+      isDefault = true;
+      settings = {
+        "browser.tabs.closeWindowWithLastTab" = false;
+      };
+    };
+  };
+}